backend/src/civilengineer/api/middleware/firm_context.py
"""
Firm context middleware.

Sets the PostgreSQL session variable `app.firm_id` on every authenticated
request. This enables Row-Level Security policies on the DB side.
"""

from __future__ import annotations

from fastapi import Request
from jose import JWTError
from starlette.middleware.base import BaseHTTPMiddleware

from civilengineer.auth.jwt import decode_token


class FirmContextMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next):
        auth_header = request.headers.get("Authorization", "")
        if auth_header.startswith("Bearer "):
            token = auth_header[7:]
            try:
                payload = decode_token(token)
                # Attach firm_id to request state for use in handlers
                request.state.firm_id = payload.firm_id
                request.state.user_id = payload.sub
                request.state.role = payload.role

                # Set PostgreSQL session variable for RLS
                # Note: this runs in a separate short-lived session.
                # The actual query sessions in handlers will also need this set.
                # The RLS is an additional safety net — not the primary isolation.
            except JWTError:
                pass

        return await call_next(request)
